fix: type error in Handsontable; move files into folders as best practice.
